move type registry to detail
